Build: Clean up internal/external build arguments, fix some compiler warnings, enable O2 optimization for SIM

+++ include/simulator/sim_gx_GlRenderer.hpp
@@ -56,7 +56,7 @@ class GlRenderer {
   int mNumChansLocation;
   int mMtxIdxALocation;
   int mPnMtxIdxEnabledLocation;
-  int mVboCapacity = 0;
+  size_t mVboCapacity = 0;
 
 
   int mBatchableDrawcalls = 0;
